@@ -14,12 +14,14 @@
     "gulp-babel": "^5.1.0",
     "gulp-streamify": "^0.0.5",
     "gulp-uglify": "^1.2.0",
+    "history": "^2.0.0-rc2",
     "jade": "^1.11.0",
     "keymirror": "^0.1.1",
     "lodash": "^3.10.0",
     "object-assign": "^3.0.0",
-    "react": "^0.13.3",
-    "react-router": "^0.13.3",
+    "react": "^0.14.6",
+    "react-dom": "^0.14.6",
+    "react-router": "^2.0.0-rc5",
     "react-tools": "^0.13.3",
     "run-parallel": "^1.1.2",
     "vinyl-source-stream": "^1.1.0",
